requirements.txt
camelot-py==1.0.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
et_xmlfile==2.0.0
fire==0.7.1
fonttools==4.61.1
lxml==6.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20251230
pillow==12.0.0
pycparser==2.23
PyMuPDF==1.26.7
pypdf==5.9.0
pypdfium2==5.2.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
shiboken6==6.10.1
six==1.17.0
tabula-py==2.10.0
tabulate==0.9.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2025.3
